refactor: move editforms from app/main to app/model
